requirements.txt
django~=5.0.1
djangorestframework~=3.14.0
PyYAML
psycopg2-binary
django-cors-headers
requests==2.32.3
